fleetai-training/requirements.txt
graphics.py~=5.0.1.post1
gym~=0.19.0
matplotlib~=3.4.3
numpy~=1.21.2
PyYAML~=5.4.1
scipy~=1.7.1
torch~=1.9.0
google-cloud-storage~=1.42.0
parse~=1.19.0
tensorboardX~=2.4
crc32c~=2.2
comet-ml~=3.15.3
